requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
python-multipart==0.0.9
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.0
fairlearn==0.10.0
google-cloud-storage==2.17.0
google-cloud-firestore==2.16.0
firebase-admin==6.5.0
python-dotenv==1.0.1
google-generativeai==0.7.2